requirements.txt
pyautogen==0.2.14
redis
python-dotenv==1.0.1
black==23.11.0
pre-commit==3.5.0
